+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 
 Changes in 0.9.X - DEV
 ======================
+- Sparse fields #515
 - write_concern not in params of Collection#remove #801
 - Better BaseDocument equality check when not saved #798
 - OperationError: Shard Keys are immutable. Tried to update id even though the document is not yet saved #771